+++ tinyui/ui/editors/preset_transfer.py
@@ -29,7 +29,6 @@ from types import MappingProxyType
 from PySide2.QtWidgets import (
     QCheckBox,
     QComboBox,
-    QFrame,
     QGridLayout,
     QHBoxLayout,
     QLabel,
@@ -46,6 +45,7 @@ from .._common import (
     CompactButton,
     UIScaler,
 )
+from ..components.list_header import ListHeader
 
 logger = logging.getLogger(__name__)
 
@@ -258,44 +258,3 @@ class PresetTransfer(BaseEditor):
                     continue
 
 
-class ListHeader(QFrame):
-    """List header"""
-
-    def __init__(self, parent, title: str, listbox: QListWidget):
-        super().__init__(parent)
-        self._parent = parent
-        self._listbox = listbox
-        self._title = title
-
-        button_selectall = CompactButton(" All ")
-        button_selectall.clicked.connect(self.button_select_all)
-
-        button_deselectall = CompactButton("None")
-        button_deselectall.clicked.connect(self.button_deselect_all)
-
-        layout = QHBoxLayout()
-        layout.setSpacing(0)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.addWidget(QLabel(f"Select {title}"))
-        layout.addWidget(button_selectall)
-        layout.addWidget(button_deselectall)
-        self.setLayout(layout)
-        self.setFrameShape(QFrame.StyledPanel)
-
-    def button_select_all(self):
-        """Select all check box"""
-        msg_text = f"Select all {self._title}s from list?"
-        if self._parent.confirm_operation(message=msg_text):
-            self.set_selection(self._listbox, True)
-
-    def button_deselect_all(self):
-        """Deselect all check box"""
-        msg_text = f"Deselect all {self._title}s from list?"
-        if self._parent.confirm_operation(message=msg_text):
-            self.set_selection(self._listbox, False)
-
-    def set_selection(self, listbox: QListWidget, checked: bool):
-        """Set check box"""
-        for row_index in range(listbox.count()):
-            item = listbox.item(row_index)
-            listbox.itemWidget(item).setChecked(checked)